chore: Move ui/route specific components to pages folder
