+++ app/javascript/pages/Keywords/SearchResultHTML.jsx
